Moved timers to own namespace and refactored
