+++ src/Outkeep.Core/Timers/SafeTimerFactory.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Threading.Tasks;
 
-namespace Outkeep
+namespace Outkeep.Timers
 {
     internal class SafeTimerFactory : ISafeTimerFactory
     {
