+++ test/Outkeep.Core.Tests/SafeTimerTests.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Logging.Abstractions;
+using Outkeep.Timers;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
